Revert byte permutes in intra-warp layout conversion (#7899)
@FrederickVu, sorry but I have to revert those 3 PRs
https://github.com/triton-lang/triton/pull/7809,
https://github.com/triton-lang/triton/pull/7825,
https://github.com/triton-lang/triton/pull/7861

There is a functional regression caused by
https://github.com/triton-lang/triton/pull/7809 but the other two PRs
have many dependencies to it so I was not able to revert it cleanly
separately and I couldn't manage to do a partial revert either.

the follow convert layout miscompiles after with this PR:
```
#blocked2 = #ttg.blocked<{sizePerThread = [1, 1], threadsPerWarp = [4, 8], warpsPerCTA = [1, 1], order = [1, 0]}>
#mma = #ttg.nvidia_mma<{versionMajor = 2, versionMinor = 0, warpsPerCTA = [1, 1], instrShape = [16, 8]}>
%4 = ttg.convert_layout %3 : tensor<16x8xf8E5M2, #mma> -> tensor<16x8xf8E5M2, #blocked2>
```

it can be reproduced on Ampere, Hopper or Blackwell GPU (I would expect
any nvidia gpu would show the problem)

I can try to get a reproducer I can share later but would be nice to
make a unit test for this convert layout in Gluon anyway.

Happy to land those back when the bug is fixed, or if you manage to
partial revert only the nvidia permute part that works too.