[AMD] Disable ptr-canonicalization for some 32-bit offset.
It is to fix https://github.com/ROCm/triton/issues/830.

Disable ptr-can for those load/stores whose base-ptr is passed via function argument with tt.ponter_range=32 attribute.
Assuming the base and offset is baseptr and ofst, respetively. The ptr-can decomposes it into uniform part (U) and non-uniform part (NonU), and let baseptr + int64(U) serve as new base pointer, and let NonU serve as new offset, the probelm is that the NonU could be out of the range of int32.