[GLUON] Fix broadcast of scalar other value for load (#7988)
When using normal `load` with scalar `other` in Gluon, the `other` value
will be broadcast using semantic's `broadcast_impl_shape`. For Gluon's
semantic, `broadcast_impl_shape` requires input to be a distributed
type. This PR changes to use semantic's `broadcast_impl_value` to
broadcast `other` (and `mask` as well) based on the pointer's the shape
and layout.