setup.py: Make `is_git_repo` more robust
The current implementation checks that `.git` is a directory. however,
if Triton is checked out in a worktree, or as a submodule, this won't be
true. There's a case for just checking that `.git` exists at all, but
using `git rev-parse --show-toplevel` seems more like the canonical way
to do it.