[BACKEND] Add linear remapping to padded shared layout (#7929)
On `GFX9`, padding alone is not enough to avoid shared memory bank
conflicts when using direct-to-LDS loads.
For example, a `global_load_lds_dwordx4` instruction loads `64 threads ×
16 bytes = 1024 bytes` at once, so padding can only be inserted at
1024-byte boundaries in LDS.

If the tensor’s rows are smaller than the transfer size, we can’t insert
padding between them, since they are loaded by the same instruction and
written contiguously to shared memory.
For instance, if each row is 256 bytes, the layout would look like:
```
[[row0], [row1], [row2], [row3], [padding], [row4], [row5], [row6], [row7], [pading], [row8] ...]
```
However, to avoid bank conflicts when reading the data, we would
actually want padding between row0 and row1, to shift elements from the
same column to different banks.

To solve this, we need to rearrange the rows in the shared memory layout
so that padding separates contiguous rows of the tensor. In this
example, by padding after every 4 rows, we could permute rows like this:

```
[[row0], [row16], [row32], [row48], [padding] [row1], [row17], [row33], [row49], [padding], [row2], [row18] ...]
```
This allows us to load the data conflict free. 

This PR adds an optional linear component to `PaddedSharedEncoding` that
allows us to permute the rows in LDS.

There will be follow-up PRs adding support for them in the stream
pipeliner and when lowering direct-to-lds loads on AMD.