[AMD] Add bypassLDS feature to StreamPipeline (#7968)
Determine if it is safe to bypass LDS for dot operands.
Normally, dot operation operands are consumed in the dot MFMA layout,
which is not coalesced. To better utilize global memory bandwidth,
operands are usually loaded in a coalesced "blocked" layout and then
rearranged through LDS.

However, certain optimizations allow dot operands to be preshuffled in
global memory. In that case, the operands can be loaded efficiently
(in a coalesced way) and consumed directly by the dot operation.
When preshuffling is used, a sequence of transpose and reshape ops
must be applied to the operand.

To verify that preshuffling was done correctly and the final layout
remains coalesced, we start from the dot MFMA layout and apply the
inverse of each transpose/reshape op (while ignoring convert_layout
ops) until we reach the load. We then inspect the resulting layout
to decide if it is coalesced enough to load directly, without needing
any further rearrangement.

TODO: getContigPerThread does not work if elements are permuted within
thread.
We need to use some utility similar to largestVectorisation() to detect
this once
load op vectorization supports in thread permutations as well.

---------

Co-authored-by: Ognjen Plavsic <plognjen@amd.com>