[WS] Use aref for TMA load pipelining and lowering (#7826)
A follow-up to https://github.com/triton-lang/triton/pull/7581 which
actually starts to use aref for TMA load. It replaces one half of
`LoadMMASpecialization`, but the code there is not removed to keep the
pass self-contained and its lit tests functional. I verified that all
tests pass if I remove the TMA code in `LoadMMASpecialization`.

`LowerAref` is updated to add
* Lowering for NVWS desc load ops
* Aref combining optimization, to coalesce barrier operations on MMA
operands into one
* Aref multi-buffering, only enabled for arefs whose producer is TMA

---------

Co-authored-by: Jeff Niu <jeffniu22@gmail.com>