Use Docker for Ubuntu x64 job
Docker allows us to maintain fine control over the environment in which
binaries are produced, which is important since we are distributing
binaries. It also keeps the build environment independent of GitHub's
deprecation schedule for images.

For example, bumping from GitHub's 20.04 runner to the 22.04 runner
also bumped the minimum required libstdc++ version, breaking binary
compatibility on systems with older toolchains. This is made trickier by
the fact that the runners come with multiple versions of GCC (including
GCC 12 in this case) and Clang always picks the newest libstdc++
available on the host.

LLVM maintains compatibility with much older compilers, so we could pick
even older images, and potentially eliminate the need to maintain
binaries for multiple distributions.