Use Docker for Ubuntu x64 job
We currently use GitHub's runners to build LLVM binaries, which are
subject to the toolchains that GitHub preinstalls on them and the
GitHub's deprecation schedule. This causes problems when we want to
build binaries for distribution, because we cannot maintain close
control of binary compatibility.

For example, bumping from GitHub's 20.04 runner to the 22.04 runner also
bumped the minimum required libstdc++ version, breaking binary
compatibility on systems with older toolchains. This is made trickier by
the fact that the runners come with multiple versions of GCC (including
GCC 12 in this case) and Clang always picks the newest libstdc++
available on the host.

Docker allows us to maintain fine control over the environment in which
binaries are produced. We can continue running old versions of
containers indefinitely and pin the exact toolchain version that is used
to build.

LLVM maintains compatibility with much older compilers, so we could pick
even older images than 22.04 and potentially eliminate the need to
maintain multiple binary distributions for Linux. But this seems like a
good first step.