[BACKEND][NVIDIA] Update ptxas_options knobs default value (#8121)
This PR sets the default value of the ptxas_options knob to be the value
of the `PTXAS_OPTIONS` environment variable.
We need this option when we want to apply extra ptxas options across a
range of Triton kernels, without changing their call sites.

# New contributor declaration

- [x] I have written a PR description following these
  [rules](https://cbea.ms/git-commit/#why-not-how).

- [x] I have run `pre-commit run --from-ref origin/main --to-ref HEAD`.

- Select one of the following.
- [x] This PR does not need a test because this is a small change of
reading ptxas options from environment variables.

- Select one of the following.
  - [x] I have not added any `lit` tests.