[BACKEND] Add bitwidth to TMEM encoding
The TMEM encoding was underspecified for `unpacked=True`. The issue here
is that we were generating the same linear layout for both
`unpacked=True` and `=False`, and then correcting for this in the
`tcgen05.ld/st` lowering.

Moving to a world where our lowerings are generic, this is not good
enough. The `unpacked=True` layout should have `log2(32 / bitwidth)` zero
bases at the beginning denoting how `32 / bitwdith - bitwidth` elements are
not defined in a 32 bit TMEM cell.

In this PR, we just add `bitwidth` to the encoding. We will modify the
linear layout and the lowerings in a follow-up.