[BACKEND] Add bitwidth to TMEM encoding (#8136)
The TMEM encoding was underspecified for `unpacked=True`. The issue here
is that we were generating the same linear layout for both
`unpacked=True` and `=False`, and then correcting for this in the
`tcgen05.ld/st` lowering.

Moving to a world where our lowerings are generic, this is not good
enough. The `unpacked=True` layout should have `log2(32 / bitwidth)`
zero
bases at the beginning denoting how `32 / bitwdith - bitwidth` elements
are
not defined in a 32 bit TMEM cell.

In this PR, we just change `unpacked` by a `colStride: int`. `colStride`
gives
you the stride along the columns. Going from the previous notation to
this one
we have:
`unpacked=True` -> `colStride=32/bitwidth`
`unpacked=False` -> `colStride=1`
This should make the semantics of the layout much cleaner.