[Backend] Fix ptxas workaround in convert_layout (#8155)
We add a fix for the workaround in #7933 that mishandled the `bitwidth <
8 && bitsPerVecElem == 8` case along with a LIT test checking that
packing 4 `i1`s to a register in a `convertLayoutOp` does not crash.

This is essentially the same as #8150, but the code in this PR will be
easier to remove once the ptxas bug is fixed. We also revert the
modifications to the vector unpacking in #7933 as they were unnecessary.

(cherry picked from commit acd72ac4060f2e38dfa0e1ac5a67d91ca7e3a76b)