[GLUON] Expose bank_conflicts and to_linear_layout
This PR exposes two a function that computes the bank_conflicts of a
distributed layout and a shared memory layout. While putting up the PR I
realised of a couple issues with the function that computes the
bank_conflicts (it does not take broadcasting into account).
I will fix those in a follow-up.