[GLUON] Fix the bank_conflict computation in the presence of shmem broadcasting
Before we forgot to model the fact that if two threads access the same
address, this address is broadcasted. We model this, and test against a
function that computes the bankconflicts bruteforcing all the memory
accesses.

We assume in the writes the hardware will just do one write per wavefront.