[WS] Rework partition representation (#8123)
`WarpSchedule` is renamed to `PartitionSet`, since it is now just a
container of created partitions. It no longer owns a mapping between ops
and partitions, which are now done exclusively by free functions like
`setPartition`, `getPartitionIds` etc which only looks at the op

Other notable changes
* Partition attributes are now a set (represented as `array<i32>`).
* The concrete instance of "root partition" is gone but the code still
uses the term to refer to ops belonging to all partitions.
* At the beginning of `PartitionLoops`, unannotated ops, including those
within if or reduce ops, are annotated with their parent partitions.