Remove CentOS 7 build
The CentOS 7 build was added back to support the release of PyTorch 2.6
(https://github.com/triton-lang/triton/pull/5158). However, PyTorch has
since dropped support for manylinux2014 and the pre-C++11 ABI:
https://dev-discuss.pytorch.org/t/pytorch-linux-wheels-switching-to-new-wheel-build-platform-manylinux-2-28-on-november-12-2024/2581

Given that the AlmaLinux 8 build already supports a relatively old
version of glibc, it is unlikely that the CentOS 7 build is being used
often. In addition, the CentOS build is tricky to use since the
libraries are built with the pre-C++11 ABI, which would need to be
propagated to the Triton build for it to link successfully.