[BACKEND][TESTS] Add `-reconcile-unrealized-casts` in "to LLVM" test
This PR adds the `-reconcile-unrealized-casts` pass to the `RUN` line of
the `tritongpu_to_llvm.mlir` test. This ensures that left-over
`unrealized_casts` ops are cleaned up, which can occur starting from
llvm/llvm-project#158298, and which would otherwise break this test.
Adding the pass is the recommended solution from the description of that
PR.

Signed-off-by: Ingo Müller <ingomueller@google.com>