[AMD][BACKEND] Coalesce direct-to-lds loads with padded encodings (#8185)
https://github.com/triton-lang/triton/pull/7929 allows us to rearrange
rows/cache line sizes chunks in LDS in addtion to padding to be able to
avoid bank conflicts. This PR adds support to rewrite the
`BlockedEncoding` to a `DistributedLinearEncoding` of the
`ttg.async_copy_global_to_local` so each warp writes a contiguous chunk
into LDS, as required by the hardware on GFX9.

Note that as long as we are just rearranging rows and do not do any
element wise swizzling the resulting LinearLayout will satisfy all
required properties of a `DistributedSharedEncoding`. The verifier of
the `PaddedSharedEncoding` checks this already and additionally the
builder of the `DistributedLinearEncoding` will trigger an assert if we
break any of the properties (surjective, moves into a single direction
per basis etc.).