Add a TritonGPU LocalMemoryBarrierOp
Membar only tracks accesses to local memory so all it needs is a thread
barrier with a fence to the local memory address space.
Before this change a GPU BarrierOp was emitted which implies much
stronger guarantees on syncronization of all address spaces and not just
local memory.

The new LocalMemoryBarrierOp is targeting only accesses between local
memory and registers to allow for better code generation.