[BACKEND] Improved swizzling when there is not enough vectorisation
In view of the corrected computation for bank conflicts discussed in
https://github.com/triton-lang/triton/pull/8200, we use the point that
anything that goes into vbasis in the same bank does not create
conflicts to expose asymmetric vectorisation whenever it would not
create more bank conflicts.

The new heuristic avoids PRMTs whenever possible on one of the
directions by choosing registers within bank 0 that are already
contiguous in the register file.

I still need to benchmark and write comprehensive tests. Will do that on
Monday.