[BACKEND] Add hook for configurable/overridable compiler pass pipeline (#8137)
Triton’s existing pass pipelines are explicitly defined in the various
extended compiler.py files that live in Triton’s backends. Currently
when we require insertion of passes either for instrumentation or for
the addition of downstream optimization and custom lowering it is
required for the compiler.py file itself to be modified.

In order to allow for more downstream configurability and as a first
step toward more custom MLIR level pass plugins, we add a hook into the
compiler stages to allow for a more configurable pass manager system
setup.

Using Python inspection routines coupled with the hook allows for more
fine grained control of things like enabling/disabling passes for
specific kernels with eventually being able to load and insert
completely out of tree ops/passes in arbitrary places in the stages
pipeline.

Co-authored with @plotfi

---------

Co-authored-by: Puyan Lotfi <puyan@puyan.org>