[ConSan] Add support for WarpSpecialization (#8189)
Introducing the support for warp specialized kernels in ConSan.
This PR changes replaces some of the ConSan auxiliary data structures
and augments others to enable tracking of read/write visibility of the
buffers between different "threads" (partitions). It introduces concepts
of TC and TMA logical threads, to model HW guarantees of intra and
inter-thread operation ordering, and verifying that these guarantees are
not being violated by the program.