[AMD] Fix assertion in CanonicalizePointers
Fix https://github.com/ROCm/triton-internal/issues/1224

The problem is an asserion arising from python/triton_kernels/bench/bench_mlp.py.

The root cause in short is that, the patten-matching code tres to use a value, say x, which is already mapped to y (by patten-matching framework), as a key to populate map (not maintained by rewrite framework). It should have used y for map index.

The problem can be depicted as following, consider two ops
  1: r = op1 ...
  2: ... op2 r, ...

The patten-matching rule tries to replace op1's result r with {a,b}
  s1: rewriter.replaceOpWithMultiple(op1, {a, b})
  s2: key = {a, b}
  s3: local-map[key] = something_awesome.

However, at this moment, b is already mapped to c. when op2 is visited. The information associated with op2.operand0 is {a, c} instead of {a, b}. If we were using {a, c} as key to lookup
the local-map, we would not find an value.

The fix is to replace s2 as:
   if (auto b2 = rewriter.getRemappedValue(b)
      b = b2;