raising error when constexpr combined with do_not_specialize (#8275)
Resolves #8273 

Previously no error message was raising when a parameter was marked as a
constexpr AND listed in the do_not_specialize list. It was just silently
treated as a compile time constant, and new specializations would be
added.

This PR checks for these cases at compile time. See added test case
explanation. Note that it will error on the call to `triton.jit` NOT the
call to execute the kernel.

# New contributor declaration
- [ ] I am not making a trivial change, such as fixing a typo in a
comment.

- [x] I have written a PR description following these
  [rules](https://cbea.ms/git-commit/#why-not-how).

- [x] I have run `pre-commit run --from-ref origin/main --to-ref HEAD`.

- Select one of the following.
  - [x] I have added tests.
    - `/test` for `lit` tests
    - `/unittest` for C++ tests
    - `/python/test` for end-to-end tests
  - [ ] This PR does not need a test because `FILL THIS IN`.

- Select one of the following.
  - [x] I have not added any `lit` tests.
- [ ] The `lit` tests I have added follow these [best
practices](https://mlir.llvm.org/getting_started/TestingGuide/#filecheck-best-practices),
including the "tests should be minimal" section. (Usually running Python
code
    and using the instructions it generates is not minimal.)

---------

Co-authored-by: ryan-yide <yide.ran@lambdal.com>