[ConSan] Deadlock Detection (#8285)
Adding deadlock detection to ConSan.
To detect deadlocks we first need to keep track of the barrier state -
we introduce `barrierState` count and phase aux structure that encodes
information about phase, init_count and current_count. We then
instrument barrier inits and arrivals - init set up init_count =
current_count = cout; phase = 0. Arrival decrements current_count and if
it hits 0, flips the phase and re-sets current_count. We treat TMA and
TC commits as barrier arrivals with count=1.
With that information we can then check if all the threads are waiting
(which constitutes a deadlock). To do that we instrument entries and
exits from barrier_wait. We introduce `waiting` aux structure, that
keeps track of waiting threads, and mapping to which barrier they are
waiting on. When issuing a wait, the thread adds itself to the
`waiting`, and checks if all the other threads are also waiting. Thread
that waits on a barrier with phase that does not match the current phase
of that barrier (info from `barrierState`) is not considered waiting.
Verified that we catch deadlocks with cases like barriers underarrival
(didn't flip the phase), overarrival (flipped the phase too many times),
etc.