[Backend] Improve interval analysis in Membar
Instead of always using the full allocation for each operation we can
calculate each interval of logical elements/offsets accessed by each
operation.
Using this mechanism we can tighten memory accesses when we are using
view operations (memdesc_index, memdesc_subslice).

The interval calculation used assumes that all elements are in shared
memory using an unswizzled row major layout. This is not as restrictive
as it sounds as allocations are usually used with the same shared layout.
The usage of different layouts for the same operation is supported by
fallback to previous membar way of using the full allocation.
It's possible to support different layouts at the same time for a single
allocation by applying the layout to each of the offsets but it would
introduce more complexity for little gain.

The improved analysis does the following:
- collect view ops, starting from usage and going up to allocation
- traverse view ops in reverse order, allocation to usage
- process view operations recursively (each adjusts just starting
  element offset)
- once all view operations are handled, generate intervals.

The processing for the various view operations is:
- MemdescSubslice: adjust starting element offset based on all strides
                   and process next view operation
- MemdescIndex (constant): adjust the starting element offset based on
                           the dim0 stride and process next view op
- MemdescIndex (dynamic): for each possible index of dim0, adjust the
                          starting element offset based on dim0 stride
                          and process the next view op

This results in fewer barriers being inserted.