[AMD] NFC: Split Loop Expansion from StreamPipeliner (#8307)
This PR refactored StreamPipeliner, with following modifications:
- exposed some structures in header
- splitted StreamPipeliner into 2 passes
- `TritonAMDGPUScheduleLoops`: to generate schedule. It keeps most of
the functionality of StreamPipeliner. In the following PR it will be
further trimmed to only schedule global load and compute.
- `TritonAMDGPUPipeline`: eventually it contains: `lowerLoops` to
allocate lds and schedule async loads, and `expandLoops` to mechanically
expand loops according to schedule. At the point of this PR, it will
only contain `expandLoops`.
- splitted `expandLoops` out of `StreamPipeliner`, into a separate
Pipeline pass
- renamed StreamPipeline.cpp to ScheduleLoops.cpp to better reflect its
purpose
- changed name of passes in lit tests accordingly

It's part of a series of PRs to refactor StreamPipeliner.

Previous PRs:
- https://github.com/triton-lang/triton/pull/8295