[Gluon] Disable constant CSE before auto layout propagation
Fixes #8229

Background is that we run the gluon inliner prior to auto layout propagation to
enable returning auto layout from a function and having different calls of the
function resolve to different layouts.

However, the inliner calls gluon canonicalize and the `GreedyPatternRewriter`
defaults to CSEing constants. This means that two distinct constants which
could otherwise resolve to different layouts may be CSEd into a single
constant and create a new conflict.

I fix this by changing the inliner to do even less canonicalization, and
only simplify control flow operations.