This change fixes https://github.com/ROCm/triton-internal/issues/1280
It is to disable poniter-canonicalization for pointer pointing to large tensors.
The large tensors refers to jit specialization for those tensor argument over 2GB.

It is disabled on the ground that is has some tricky bugs. We are trying to come up a better approach that address several conflicting performance aspects.