[AMD] disable pointer-canonicalization for large-tensor (#8359)
This commit disables pointer-canonicalization for pointer pointing to large
tensors. The large tensors refers to JIT specialization for those tensor
argument over 2GB.

It is disabled on the ground that is has some tricky bugs. We are trying
to come up a better approach that address several conflicting
performance aspects.

---------

Co-authored-by: Shuxin Yang <Shuxin.Yang@gmail.com>