Get MLIRContext from `newOp`, not the deleted `load` (#8373)
The `load` op is erased in `convertDistributedOpEncoding` so we should
get the MLIR context from the `newOp` when setting op attributes.
Interestingly, I found this issue only because the
`loop-pipeline-hip.mlir` lit test crashed with bad access on my Mac
laptop (ARM). On Linux/x86 the test runs fine w/ and w/out this fix.