[LAYOUTS] Generate distributed layouts for `tcgen05.ld/st` generically
We move the previous handwritten logic to infer good distributed layout
for TMEM layouts to a generic logic. This proves to be more robust than
the previous one, as we see in the lit tests, where we are able to get
full vectorisation in many cases that we didn't before.

Writing this genrically also allows us to add support for the two remaining
tcgen05.ld/st instructions that were missing.

We expose this function in gluon as well