[AMD] Refactor FP conversion mode setting (#8351)
In the current implementation we reset mode register every time when we
perform FP conversion to FP8 data type. We modify F16_OVFL flag which
also effects clamping during conversions of the FP16 data type. In fact,
the flag should be inserted only one (e.g., at the beginning of a
kernel). This PR addresses this issue. It moves the manipulation with
the mode register to a dedicated function which gets initialized with an
`AMD::ISAFamily` instance. Note, the the layout of bits in mode register
may vary from architecture to architecture.