Use system toolchain for AlmaLinux build
The AlmaLinux build currently builds against libstdc++ from GCC 14,
because installing clang also pulls in GCC 14, and clang automatically
picks the newest libstdc++ on the system.

Instead, using the system toolchain (which is GCC 8), should link
against the corresponding libstdc++. This improves the portability of
the generated binaries. The AlmaLinux binaries should now be compatible
with a superset of the systems that the Ubuntu binaries are compatible
with.

A subsequent change can then make this the default set of binaries used
on Linux regardless of glibc version.

(I assume these have to be separate changes for the CI to pass)