Remove CentOS 7 build (#8191)
The CentOS 7 build was added back to support the release of PyTorch 2.6
(https://github.com/triton-lang/triton/pull/5158). However, PyTorch has
since dropped support for manylinux2014 and the pre-C++11 ABI:

https://dev-discuss.pytorch.org/t/pytorch-linux-wheels-switching-to-new-wheel-build-platform-manylinux-2-28-on-november-12-2024/2581

Given that the AlmaLinux 8 build already supports a relatively old
version of glibc, it is unlikely that the CentOS 7 build is being used
often. In addition, the CentOS build is tricky to use since the
libraries are built with the pre-C++11 ABI, which would need to be
propagated to the Triton build for it to link successfully.

# New contributor declaration
- [x] I am not making a trivial change, such as fixing a typo in a
comment.

- [x] I have written a PR description following these
  [rules](https://cbea.ms/git-commit/#why-not-how).

- [x] I have run `pre-commit run --from-ref origin/main --to-ref HEAD`.

- Select one of the following.
  - [ ] I have added tests.
    - `/test` for `lit` tests
    - `/unittest` for C++ tests
    - `/python/test` for end-to-end tests
  - [x] This PR does not need a test because it is a CI change.

- Select one of the following.
  - [x] I have not added any `lit` tests.
- [ ] The `lit` tests I have added follow these [best
practices](https://mlir.llvm.org/getting_started/TestingGuide/#filecheck-best-practices),
including the "tests should be minimal" section. (Usually running Python
code
    and using the instructions it generates is not minimal.)