[AMD] Lift ds_read_tr restrictions as they are not needed
The only restrictions that are needed to be able to lower ds_read_tr are
based on vectorization.
The ds_read_tr instruction works just by shuffling elements read by each
lane with the elements read by the group of lanes. This is now better
reflected in how the generation of ldsTransLayout is performed.