[AMD] Use v_dot for bf16 multiplication on gfx11/gfx12
RDNA does not support bf16 multiplication, so we currently upcast to
fp32, multiply and downcast to bf16 again. We can work around the
expensive downcasting step by using the V_DOT2_BF16_BF16 instruction to
calculate res = a*b + 0*0 + 0 which allows us to do round to nearest
even in hardware. The result is bit for bit identical to the current
implementation, but significantly faster.