[LAYOUTS] Generate distributed layouts for `tcgen05.ld/st` generically (#8421)
We move the previous handwritten logic to infer good distributed layout
for TMEM layouts to a generic logic. This proves to be more robust than
the previous one, as we see in the lit tests, where we are able to get
full vectorisation in many cases that we didn't before.

Writing this generically also allows us to add support for the two
remaining
tcgen05.ld/st instructions that were missing.

We align the verifier and the lowering as to now error out during
verification
if and only if we would not be able to lower the given layout.

We expose this function in gluon as well and kill the duplicated logic
that we
had in gluon in favour of the generic logic.

**There is just one semantic change** (the rest is generalisation /
strengthening):
We now generate distributed layouts that generate the fully vectorised
load/store
instructions (i.e. they would generate just one load/store instruction
to load all the
registers). As we see in the `_blackwell.mlir` cases, this did not use
to be the case.
That being said, in some cases the previous heuristics allowed to
`tt.split` the tensor
along the second dimension, while now this may not be the case. If one
wants to perform
this splitting game, they need to modify a bit the layout as we do when
calling `32x32b_splitn`
in the test `test_tmem_subslice_block_m_64`.
This change goes in line with the rest of the heuristics, where we have
full vectorisation by
default, and we expose other layouts for other use cases like
`splitLongM`.