[DRAFT][AMD] fix ptr-canonicalization pass segmentation fault
This change fix https://github.com/ROCm/triton-internal/issues/1326 and https://github.com/ROCm/frameworks-internal/issues/14125

This is a very old bug. The problem can be explained by following simple statement, where the fatPtrs is a simple wrapper of a hashmap.

  fatPtrs[toKey] = fatPtrs.at(fromKey)

time 0: RHS's at() function return a reference to an existing element.
time 1: There is no element associated with toKey, so the LHS needs to resize its internal bucket in order to accommodate the new entry. When resize, it invalidates the reference it took in the step 1

The fix is to change the map's value type from an aggregate type ValueT to std::unique_ptr<ValueT>. With this change, when map resizes, while unique_ptr may change, the raw pointer it maintains is unchanged, and hence we can get a "sticky" pointer for the data it maps to.