[AMD] Add gfx1250 support for ds_read_tr (#8461)
In gfx1250 the instruction has different size based on datatype (b128
for b16 types and b64 for b8 types).
Currently the ROCDL transpose instructions aren't upstream so this is
using LLVM intrinsics.
Once the ROCDL instructions are available we will use those.

As part of the change I've refactored the ds_read_tr parameter setup