Fix AxisInfo handling of PoisonOp producing MemDesc
AxisInfo analysis currently retrieves the rank from any `ShapedType`
producing `PoisonOp`. This is a problem if the `PoisonOp` actually
produces a `MemDesc`, since the value produced by the `PoisonOp` may
flow into the same value as some other `MemDesc` producing operation,
which will have been assigned the "pessimistic state" and have rank 1.

When we attempt to join the two, the ranks will not match, potentially
resulting in a crash.