[AMD] Expose buffer ops small tensor analysis to Python (#8471)
This allows to test both the functionality and effectiveness of the
additional small tensor argument to Converting to BufferOps. To enable
this testing I added a flag to `add_convert_to_buffer_ops` since I
assume there may be some concern about enabling this by default.

My expectation is that long-term this will not be a flag, but this would
also be good to have to simplify debugging once we want to enable this
analysis by default.